fix(customer_app): C-4 Stage B boundary conversions — int cents ↔ double SAR
After cherry-picking 2236b94 onto fix/zatca-queue-drift HEAD, the app failed
to build with 7 test failures + 3 compile errors. Root cause: customer_app
was branched before C-4 Stage B made Product.price/costPrice int cents (from
double SAR). 5 boundary sites needed conversion:

Ingress (Supabase INTEGER cents → Product int cents):
- products_datasource.dart:88/89: .toDouble() → .toInt()

Product → Cart (int cents → double SAR):
- cart_provider.dart:95: product.price → product.price / 100.0

Display (int cents → double SAR):
- search_screen.dart:152
- catalog_screen.dart:432
- product_detail_screen.dart:101

All three display sites: product.price.toStringAsFixed(2)
                         → (product.price / 100.0).toStringAsFixed(2)

Pattern: cents stored, doubles for UI math — same boundary design used in
edit_price_screen fix (72d04e1) and the rest of Stage B.

Verification:
- flutter analyze lib/: 0 errors (2 pre-existing unrelated info/warning)
- flutter test: 136/136 ✓ (baseline restored from +101 -7)

+++ customer_app/lib/features/catalog/data/products_datasource.dart
@@ -85,8 +85,9 @@ class ProductsDatasource {
       name: row['name'] as String,
       sku: row['sku'] as String?,
       barcode: row['barcode'] as String?,
-      price: (row['price'] as num).toDouble(),
-      costPrice: (row['cost_price'] as num?)?.toDouble(),
+      // C-4 Stage B: server sends int cents (post v71); Product.price is int.
+      price: (row['price'] as num).toInt(),
+      costPrice: (row['cost_price'] as num?)?.toInt(),
       stockQty: (row['stock_qty'] as num?)?.toDouble() ?? 0,
       minQty: (row['min_qty'] as num?)?.toDouble() ?? 1,
       unit: row['unit'] as String?,
